funcao-geradora-url/requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==3.*
google-cloud-storage==2.* 